first shot at reorganizing library
I am not sure if this structure is correct - it could be the case that we do not want both
allocators.hpp and allocators.cpp, but instead just the hpp. And then I am also
wondering how we decide to put something in the include folder vs. alongside
a cpp file like in this case?
