+++ source/parser/x86_64/x86_64.cpp
@@ -14,129 +14,13 @@
 #include "Function.h"
 #include "Symtab.h"
 #include "Type.h"
+#include "allocators.hpp"
 #include "smeagle/parameter.h"
+#include "type_checker.hpp"
 
 namespace smeagle::x86_64 {
 
   namespace st = Dyninst::SymtabAPI;
-
-  // A register class for AMD64 is defined on page 16 of the System V abi pdf
-  enum RegisterClass {
-    INTEGER,      // Integer types that fit into one of the general purpose registers
-    SSE,          // Types that fit into an SSE register
-    SSEUP,        // ^.. and can ve passed and returned in he most significant half of it
-    X87,          // Types that will be returned via the x87 FPU
-    X87UP,        // ^
-    COMPLEX_X87,  // Types that will be returned via the x87 FPU
-    NO_CLASS,     // Initalizer in the algorithms, used for padding and empty
-                  // tructures and unions
-    MEMORY        // Types that will be passed and returned in memory via the stack
-  };
-
-  // A FramebaseAllocator keeps track of framebase index
-  class FramebaseAllocator {
-  private:
-    int nextMultipleEight(int number);
-    int framebase = 8;
-    void updateFramebaseFromType(st::Type *paramType);
-
-  public:
-    std::string nextFramebaseFromType(st::Type *paramType);
-  };
-
-  // Get a framebase for a variable based on stack location and type
-  // Framebase values must be 8 byte aligned.
-  void FramebaseAllocator::updateFramebaseFromType(st::Type *paramType) {
-    framebase += nextMultipleEight(paramType->getSize());
-  }
-
-  // Get the next greater multiple of 8
-  int FramebaseAllocator::nextMultipleEight(int number) { return ((number + 7) & (-8)); }
-
-  // Get the next framebase
-  std::string FramebaseAllocator::nextFramebaseFromType(st::Type *paramType) {
-    std::string framebaseStr = "framebase+" + std::to_string(framebase);
-
-    // Update the framebase for the next parameter based on the type
-    updateFramebaseFromType(paramType);
-
-    return framebaseStr;
-  }
-
-  // A RegisterAllocator can provide the next register location
-  class RegisterAllocator {
-  public:
-    RegisterAllocator();
-    std::string getRegisterString(RegisterClass regClass, st::Type *paramType);
-    std::string getRegistersString(std::pair<RegisterClass, RegisterClass> regClasses,
-                                   st::Type *paramType);
-
-  private:
-    FramebaseAllocator fallocator;
-    std::stack<std::string> intRegisters{{"%r9", "%r8", "%rcx", "%rdx", "%rsi", "%rdi"}};
-    std::stack<std::string> sseRegisters;
-    std::optional<std::string> getNextIntRegister();
-    std::optional<std::string> getNextSseRegister();
-    int framebase = 8;
-  };
-
-  // Constructor
-  RegisterAllocator::RegisterAllocator() {
-    // Populate the sse register stack
-    for (int i = 1; i < 8; ++i) {
-      sseRegisters.push("%xmm" + std::to_string(i));
-    }
-  }
-
-  // Get the next available integer register
-  std::optional<std::string> RegisterAllocator::getNextIntRegister() {
-    // If we are empty, return stack
-    if (intRegisters.empty()) return {};
-    std::string regString = intRegisters.top();
-    intRegisters.pop();
-    return regString;
-  }
-
-  // Get the next available integer register
-  std::optional<std::string> RegisterAllocator::getNextSseRegister() {
-    // If we are empty, return stack
-    if (sseRegisters.empty()) return {};
-    std::string regString = sseRegisters.top();
-    sseRegisters.pop();
-    return regString;
-  }
-
-  // Given two registers, return one combined string
-  std::string RegisterAllocator::getRegistersString(
-      std::pair<RegisterClass, RegisterClass> regClasses, st::Type *paramType) {
-    std::string locA = this->getRegisterString(regClasses.first, paramType);
-    std::string locB = this->getRegisterString(regClasses.second, paramType);
-
-    // If B is empty (NO_CLASS) then return A
-    if (locB == "") {
-      return locA;
-    }
-    return locA + "|" + locB;
-  }
-
-  // Get a string location from a register class
-  std::string RegisterAllocator::getRegisterString(RegisterClass regClass, st::Type *paramType) {
-    std::optional<std::string> regString;
-
-    // If the class is memory, pass the argument on the stack
-    if (regClass == RegisterClass::NO_CLASS) regString = "";
-    if (regClass == RegisterClass::SSE) regString = this->getNextSseRegister();
-    if (regClass == RegisterClass::INTEGER) regString = this->getNextIntRegister();
-    if (regClass == RegisterClass::MEMORY) regString = std::nullopt;
-
-    // If we don't have a value, we need a framebase
-    if (!regString.has_value()) {
-      regString = fallocator.nextFramebaseFromType(paramType);
-    }
-
-    // If we've run out of registers we get to this point
-    return regString.value();
-  }
 
   // Given a symbol, get a string representation of its type
   std::string getStringSymbolType(st::Symbol *symbol) {
@@ -184,20 +68,6 @@ namespace smeagle::x86_64 {
     return sname;
   }
 
-  // Return true if the dataClass is a pointer type
-  static bool is_pointer(st::dataClass dc) { return dc == st::dataPointer; }
-
-  // Return true if the dataClass is a reference
-  static bool is_ref(st::dataClass dc) { return dc == st::dataReference; }
-
-  // Return true if dataClass is reference or pointer
-  static bool is_indirect(st::dataClass dc) {
-    return dc == st::dataPointer || dc == st::dataReference;
-  }
-
-  // Return true if dataClass is a primitive type
-  static bool is_primitive(st::dataClass dc) { return dc == st::dataEnum || dc == st::dataScalar; }
-
   // Dereference a pointer or reference
   static st::Type *deref(st::Type *t) {
     if (is_pointer(t->getDataClass())) {
@@ -208,8 +78,6 @@ namespace smeagle::x86_64 {
     }
     return t;
   }
-
-  static bool is_typedef(st::dataClass dc) { return dc == st::dataTypedef; }
 
   // Get a location offset for a variable
   // This function is not used because LocationLists are not reliable
