+++ test/source/allocation.cpp
@@ -8,24 +8,24 @@
 #include "smeagle/smeagle.h"
 
 auto get_one(smeagle::Corpus const& corpus, char const* name) {
-	  auto funcs = corpus.getFunctions();
+  auto funcs = corpus.getFunctions();
 
-	  // We only want the "foo" function
-	  funcs.erase(std::remove_if(funcs.begin(), funcs.end(),
-	                             [name](smeagle::abi_description const& d) {
-	                               return d.function_name.find(name) == std::string::npos;
-	                             }),
-	              funcs.end());
-	  return funcs;
+  // We only want the "foo" function
+  funcs.erase(std::remove_if(funcs.begin(), funcs.end(),
+                             [name](smeagle::abi_description const& d) {
+                               return d.function_name.find(name) == std::string::npos;
+                             }),
+              funcs.end());
+  return funcs;
 }
 
 TEST_CASE("Register Allocation") {
   auto corpus = smeagle::Smeagle("liballocation.so").parse();
 
   SUBCASE("one int") {
-	  auto funcs = get_one(corpus, "f1");
-	  REQUIRE(funcs.size() == 1);
-	  auto const& parameters = funcs[0].parameters;
-	  CHECK(parameters[0].location == "%rdi");
+    auto funcs = get_one(corpus, "f1");
+    REQUIRE(funcs.size() == 1);
+    auto const& parameters = funcs[0].parameters;
+    CHECK(parameters[0].location == "%rdi");
   }
 }
