+++ test/source/libs/allocation.cpp
@@ -1,4 +1,3 @@
 // Functions to test register allocation
 
 void f1(int x) {}
-
